requirements.txt
fastapi
manage-fastapi
uvicorn
FastAPI-SQLAlchemy
pydantic==1.10.4
python-dotenv==1.0.0
mysqlclient
PyMySQL
